requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
sentinelhub==3.10.1
llama-cpp-python==0.2.23
pydantic==2.5.0
